backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.5.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
clarifai==9.7.1
clarifai-grpc==9.7.3
cohere==4.20.2
colorama==0.4.6
dataclasses-json==0.5.14
dill==0.3.7
Django==4.1.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.1.3
fastavro==1.8.2
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.57.0
gunicorn==20.1.0
h11==0.14.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
langchain==0.0.268
langsmith==0.0.25
manifest-ml==0.0.1
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nlpcloud==1.1.44
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openlm==0.0.5
outcome==1.2.0
packaging==23.1
Pillow==10.0.0
protobuf==4.24.1
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
PyJWT==2.6.0
PySocks==1.7.1
pytesseract==0.3.10
python-dotenv==0.21.0
python-rapidjson==1.10
pytz==2022.4
PyVirtualDisplay==3.0
PyYAML==6.0.1
redis==5.0.0
regex==2023.8.8
requests==2.31.0
safetensors==0.3.2
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
sqlitedict==2.1.0
sqlparse==0.4.3
sympy==1.12
tenacity==8.2.3
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.31.0
trio==0.22.2
trio-websocket==0.10.3
tritonclient==2.34.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2022.4
urllib3==2.0.4
wsproto==1.2.0
yarl==1.9.2
zipp==3.16.2
